requirements.txt
Flask==3.0.0
ImageHash==4.3.1
Pillow==10.0.0
numpy==1.26.0
requests==2.31.0
python-dotenv==1.0.0